07-01/pyproject.toml
[project]
name = "07-01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckduckgo-search>=7.2.1",
    "google-generativeai>=0.8.3",
    "phidata>=2.7.7",
    "yfinance>=0.2.51",
]
